+++ test-suite/executable-benchmark/arrow-crossing/src.hpp
@@ -1,0 +1,6 @@
+#pragma once
+#include "mlc_arrow.hpp"
+
+inline mlc::ArrowTable idCpp(const mlc::ArrowTable& t) {
+    return t;
+}